Fix exploration, update score

+++ src/current/States/AttackCat.java
@@ -43,7 +43,7 @@ public class AttackCat extends State {
                 PathFinding.resetScores();
                 PathFinding.modificatorOrientation(awayFromCat);
                 // Boost directions that get us out of cat view
-                int[] escapeScores = new int[9];
+                long[] escapeScores = new long[9];
                 for (Direction dir : Direction.values()) {
                     if (dir != Direction.CENTER) {
                         MapLocation nextLoc = myLoc.add(dir);
@@ -79,7 +79,7 @@ public class AttackCat extends State {
             PathFinding.resetScores();
             PathFinding.modificatorOrientation(awayFromCat);
             // Boost directions that get us out of cat view
-            int[] escapeScores = new int[9];
+            long[] escapeScores = new long[9];
             for (Direction dir : Direction.values()) {
                 if (dir != Direction.CENTER) {
                     MapLocation nextLoc = myLoc.add(dir);
